refactor code halaman infografis dan perbaikan responsive halaman data
