+++ src/app/data-infografis/infografis/dataInfografis.tsx
@@ -1,0 +1,350 @@
+type ImageType = {
+  url: string;
+  alt: string;
+  prov: string;
+};
+
+export type DataInfografisType = {
+  tag: string;
+  title: string;
+  image: ImageType[];
+};
+
+export const dataInfografis = [
+  {
+    tag: "pertumbuhan-ekonomi",
+    title: "Pertumbuhan Ekonomi",
+    image: [
+      {
+        url: "/assets/infografis/ipm/aceh.jpg",
+        alt: "IPM Aceh",
+        prov: "Provinsi Aceh",
+      },
+      {
+        url: "/assets/infografis/ipm/sumut.jpg",
+        alt: "IPM Sumut",
+        prov: "Provinsi Sumatera Utara",
+      },
+      {
+        url: "/assets/infografis/ipm/sumbar.png",
+        alt: "IPM Sumbar",
+        prov: "Provinsi Sumatera Barat",
+      },
+      {
+        url: "/assets/infografis/ipm/riau.jpg",
+        alt: "IPM Riau",
+        prov: "Provinsi Riau",
+      },
+      {
+        url: "/assets/infografis/ipm/jambi.jpg",
+        alt: "IPM Jambi",
+        prov: "Provinsi Jambi",
+      },
+      {
+        url: "/assets/infografis/ipm/sumsel.jpg",
+        alt: "IPM Sumsel",
+        prov: "Provinsi Sumatera Selatan",
+      },
+      {
+        url: "/assets/infografis/ipm/bengkulu.png",
+        alt: "IPM Bengkulu",
+        prov: "Provinsi Bengkulu",
+      },
+      {
+        url: "/assets/infografis/ipm/lampung.jpg",
+        alt: "IPM Lampung",
+        prov: "Provinsi Lampung",
+      },
+      {
+        url: "/assets/infografis/ipm/babel.jpg",
+        alt: "IPM Babel",
+        prov: "Provinsi Bangka Belitung",
+      },
+      {
+        url: "/assets/infografis/ipm/kepri.png",
+        alt: "IPM Kepri",
+        prov: "Provinsi Kepulauan Riau",
+      },
+    ],
+  },
+  {
+    tag: "kemiskinan",
+    title: "Kemiskinan",
+    image: [
+      {
+        url: "/assets/infografis/ipm/aceh.jpg",
+        alt: "IPM Aceh",
+        prov: "Provinsi Aceh",
+      },
+      {
+        url: "/assets/infografis/ipm/sumut.jpg",
+        alt: "IPM Sumut",
+        prov: "Provinsi Sumatera Utara",
+      },
+      {
+        url: "/assets/infografis/ipm/sumbar.png",
+        alt: "IPM Sumbar",
+        prov: "Provinsi Sumatera Barat",
+      },
+      {
+        url: "/assets/infografis/ipm/riau.jpg",
+        alt: "IPM Riau",
+        prov: "Provinsi Riau",
+      },
+      {
+        url: "/assets/infografis/ipm/jambi.jpg",
+        alt: "IPM Jambi",
+        prov: "Provinsi Jambi",
+      },
+      {
+        url: "/assets/infografis/ipm/sumsel.jpg",
+        alt: "IPM Sumsel",
+        prov: "Provinsi Sumatera Selatan",
+      },
+      {
+        url: "/assets/infografis/ipm/bengkulu.png",
+        alt: "IPM Bengkulu",
+        prov: "Provinsi Bengkulu",
+      },
+      {
+        url: "/assets/infografis/ipm/lampung.jpg",
+        alt: "IPM Lampung",
+        prov: "Provinsi Lampung",
+      },
+      {
+        url: "/assets/infografis/ipm/babel.jpg",
+        alt: "IPM Babel",
+        prov: "Provinsi Bangka Belitung",
+      },
+      {
+        url: "/assets/infografis/ipm/kepri.png",
+        alt: "IPM Kepri",
+        prov: "Provinsi Kepulauan Riau",
+      },
+    ],
+  },
+  {
+    tag: "ketenagakerjaan",
+    title: "Ketenagakerjaan",
+    image: [
+      {
+        url: "/assets/infografis/ipm/aceh.jpg",
+        alt: "IPM Aceh",
+        prov: "Provinsi Aceh",
+      },
+      {
+        url: "/assets/infografis/ipm/sumut.jpg",
+        alt: "IPM Sumut",
+        prov: "Provinsi Sumatera Utara",
+      },
+      {
+        url: "/assets/infografis/ipm/sumbar.png",
+        alt: "IPM Sumbar",
+        prov: "Provinsi Sumatera Barat",
+      },
+      {
+        url: "/assets/infografis/ipm/riau.jpg",
+        alt: "IPM Riau",
+        prov: "Provinsi Riau",
+      },
+      {
+        url: "/assets/infografis/ipm/jambi.jpg",
+        alt: "IPM Jambi",
+        prov: "Provinsi Jambi",
+      },
+      {
+        url: "/assets/infografis/ipm/sumsel.jpg",
+        alt: "IPM Sumsel",
+        prov: "Provinsi Sumatera Selatan",
+      },
+      {
+        url: "/assets/infografis/ipm/bengkulu.png",
+        alt: "IPM Bengkulu",
+        prov: "Provinsi Bengkulu",
+      },
+      {
+        url: "/assets/infografis/ipm/lampung.jpg",
+        alt: "IPM Lampung",
+        prov: "Provinsi Lampung",
+      },
+      {
+        url: "/assets/infografis/ipm/babel.jpg",
+        alt: "IPM Babel",
+        prov: "Provinsi Bangka Belitung",
+      },
+      {
+        url: "/assets/infografis/ipm/kepri.png",
+        alt: "IPM Kepri",
+        prov: "Provinsi Kepulauan Riau",
+      },
+    ],
+  },
+  {
+    tag: "ipm",
+    title: "Indeks Pembangunan Manusia (IPM)",
+    image: [
+      {
+        url: "/assets/infografis/ipm/aceh.jpg",
+        alt: "IPM Aceh",
+        prov: "Provinsi Aceh",
+      },
+      {
+        url: "/assets/infografis/ipm/sumut.jpg",
+        alt: "IPM Sumut",
+        prov: "Provinsi Sumatera Utara",
+      },
+      {
+        url: "/assets/infografis/ipm/sumbar.png",
+        alt: "IPM Sumbar",
+        prov: "Provinsi Sumatera Barat",
+      },
+      {
+        url: "/assets/infografis/ipm/riau.jpg",
+        alt: "IPM Riau",
+        prov: "Provinsi Riau",
+      },
+      {
+        url: "/assets/infografis/ipm/jambi.jpg",
+        alt: "IPM Jambi",
+        prov: "Provinsi Jambi",
+      },
+      {
+        url: "/assets/infografis/ipm/sumsel.jpg",
+        alt: "IPM Sumsel",
+        prov: "Provinsi Sumatera Selatan",
+      },
+      {
+        url: "/assets/infografis/ipm/bengkulu.png",
+        alt: "IPM Bengkulu",
+        prov: "Provinsi Bengkulu",
+      },
+      {
+        url: "/assets/infografis/ipm/lampung.jpg",
+        alt: "IPM Lampung",
+        prov: "Provinsi Lampung",
+      },
+      {
+        url: "/assets/infografis/ipm/babel.jpg",
+        alt: "IPM Babel",
+        prov: "Provinsi Bangka Belitung",
+      },
+      {
+        url: "/assets/infografis/ipm/kepri.png",
+        alt: "IPM Kepri",
+        prov: "Provinsi Kepulauan Riau",
+      },
+    ],
+  },
+  {
+    tag: "gini-ratio",
+    title: "Gini Ratio",
+    image: [
+      {
+        url: "/assets/infografis/ipm/aceh.jpg",
+        alt: "IPM Aceh",
+        prov: "Provinsi Aceh",
+      },
+      {
+        url: "/assets/infografis/ipm/sumut.jpg",
+        alt: "IPM Sumut",
+        prov: "Provinsi Sumatera Utara",
+      },
+      {
+        url: "/assets/infografis/ipm/sumbar.png",
+        alt: "IPM Sumbar",
+        prov: "Provinsi Sumatera Barat",
+      },
+      {
+        url: "/assets/infografis/ipm/riau.jpg",
+        alt: "IPM Riau",
+        prov: "Provinsi Riau",
+      },
+      {
+        url: "/assets/infografis/ipm/jambi.jpg",
+        alt: "IPM Jambi",
+        prov: "Provinsi Jambi",
+      },
+      {
+        url: "/assets/infografis/ipm/sumsel.jpg",
+        alt: "IPM Sumsel",
+        prov: "Provinsi Sumatera Selatan",
+      },
+      {
+        url: "/assets/infografis/ipm/bengkulu.png",
+        alt: "IPM Bengkulu",
+        prov: "Provinsi Bengkulu",
+      },
+      {
+        url: "/assets/infografis/ipm/lampung.jpg",
+        alt: "IPM Lampung",
+        prov: "Provinsi Lampung",
+      },
+      {
+        url: "/assets/infografis/ipm/babel.jpg",
+        alt: "IPM Babel",
+        prov: "Provinsi Bangka Belitung",
+      },
+      {
+        url: "/assets/infografis/ipm/kepri.png",
+        alt: "IPM Kepri",
+        prov: "Provinsi Kepulauan Riau",
+      },
+    ],
+  },
+  {
+    tag: "pdrb",
+    title: "PDRB Per Kapita",
+    image: [
+      {
+        url: "/assets/infografis/ipm/aceh.jpg",
+        alt: "IPM Aceh",
+        prov: "Provinsi Aceh",
+      },
+      {
+        url: "/assets/infografis/ipm/sumut.jpg",
+        alt: "IPM Sumut",
+        prov: "Provinsi Sumatera Utara",
+      },
+      {
+        url: "/assets/infografis/ipm/sumbar.png",
+        alt: "IPM Sumbar",
+        prov: "Provinsi Sumatera Barat",
+      },
+      {
+        url: "/assets/infografis/ipm/riau.jpg",
+        alt: "IPM Riau",
+        prov: "Provinsi Riau",
+      },
+      {
+        url: "/assets/infografis/ipm/jambi.jpg",
+        alt: "IPM Jambi",
+        prov: "Provinsi Jambi",
+      },
+      {
+        url: "/assets/infografis/ipm/sumsel.jpg",
+        alt: "IPM Sumsel",
+        prov: "Provinsi Sumatera Selatan",
+      },
+      {
+        url: "/assets/infografis/ipm/bengkulu.png",
+        alt: "IPM Bengkulu",
+        prov: "Provinsi Bengkulu",
+      },
+      {
+        url: "/assets/infografis/ipm/lampung.jpg",
+        alt: "IPM Lampung",
+        prov: "Provinsi Lampung",
+      },
+      {
+        url: "/assets/infografis/ipm/babel.jpg",
+        alt: "IPM Babel",
+        prov: "Provinsi Bangka Belitung",
+      },
+      {
+        url: "/assets/infografis/ipm/kepri.png",
+        alt: "IPM Kepri",
+        prov: "Provinsi Kepulauan Riau",
+      },
+    ],
+  },
+];
