--- src/app/data-infografis/infografis/ketenagakerjaan/page.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function pages() {
-  return <div>Ketenagakerjaan</div>;
-}
