--- src/app/data-infografis/infografis/pdrb/page.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function pages() {
-  return <div>pdrb</div>;
-}
