--- src/app/data-infografis/infografis/pertumbuhan-ekonomi/dataInfografis.tsx
@@ -1,56 +1,0 @@
-export const dataInfografis = {
-  tag: "pertumbuhan-ekonomi",
-  title: "Pertumbuhan Ekonomi",
-  image: [
-    {
-      url: "/assets/infografis/ipm/aceh.jpg",
-      alt: "IPM Aceh",
-      prov: "Provinsi Aceh",
-    },
-    {
-      url: "/assets/infografis/ipm/sumut.jpg",
-      alt: "IPM Sumut",
-      prov: "Provinsi Sumatera Utara",
-    },
-    {
-      url: "/assets/infografis/ipm/sumbar.png",
-      alt: "IPM Sumbar",
-      prov: "Provinsi Sumatera Barat",
-    },
-    {
-      url: "/assets/infografis/ipm/riau.jpg",
-      alt: "IPM Riau",
-      prov: "Provinsi Riau",
-    },
-    {
-      url: "/assets/infografis/ipm/jambi.jpg",
-      alt: "IPM Jambi",
-      prov: "Provinsi Jambi",
-    },
-    {
-      url: "/assets/infografis/ipm/sumsel.jpg",
-      alt: "IPM Sumsel",
-      prov: "Provinsi Sumatera Selatan",
-    },
-    {
-      url: "/assets/infografis/ipm/bengkulu.png",
-      alt: "IPM Bengkulu",
-      prov: "Provinsi Bengkulu",
-    },
-    {
-      url: "/assets/infografis/ipm/lampung.jpg",
-      alt: "IPM Lampung",
-      prov: "Provinsi Lampung",
-    },
-    {
-      url: "/assets/infografis/ipm/babel.jpg",
-      alt: "IPM Babel",
-      prov: "Provinsi Bangka Belitung",
-    },
-    {
-      url: "/assets/infografis/ipm/kepri.png",
-      alt: "IPM Kepri",
-      prov: "Provinsi Kepulauan Riau",
-    },
-  ],
-};
